Remove extraneous `!` from test commands
The `expect` and `assert` functions both accept `!` as their first
argument, which negates the test. However, those functions were
putting the `!` into the executed command, which is not correct
shell syntax and can cause exrtraneous errors. The `!` is now
`shift`ed out of the args and only the tested command is run.